error handlers for controllers

+++ app/Http/Controllers/ProfitgoalController.php
@@ -22,13 +22,20 @@ class ProfitgoalController extends Controller
 }
 
 public function getProfitgoal(Request $request, $id ){
-    $profitgoal = Profitgoal::where('id',$id)->get();
+    try{
+    $profitgoal = Profitgoal::where('id',$id)->get()->firstOrFail();
+     }catch(\Exception $exception){
 
     return response()->json([
-        'message' => $profitgoal,
+        'message' => 'the profitgoal is not found.',
     ]);
+}
+return response()->json([
+ 'message' => $profitgoal,
+  ]);
 
 }
+
 
 public function getProfitgoalAll(Request $request){
     $profitgoal = Profitgoal::get();
@@ -44,20 +51,38 @@ public function getProfitgoalAll(Request $request){
 public function editProfitgoal(Request $request, $id ){
     $profitgoal = Profitgoal::find($id);
     $inputs=$request->except('_method');
-    $profitgoal->update($inputs);
+    if ($profitgoal){
+         $profitgoal->update($inputs);
     return  response()->json([
         'message' =>'Profitgoal edited successtully',
         'Profitgoal' =>$profitgoal,
     ]);
 
+    }else{
+          return  response()->json([
+        'message' =>'The profit goal does not exist.',
+      
+    ]);
+
+    }
+   
     }
 
     public function deleteProfitgoal(Request $request, $id){
         $profitgoal = Profitgoal::find($id);
-        $profitgoal->delete();
-        return response()->json([
+        if ($profitgoal){
+             $profitgoal->delete();
+                 return response()->json([
             'message' =>'Profitgoal deleted successtully',
-            'Profitgoal' =>$profitgoal,
         ]);
+
+        }else{
+              return  response()->json([
+        'message' =>'The profit goal does not exist.',
+      
+    ]);
+
+        }
     }
 }
+
